Added pragmas to disable unnecessary warning; changed type of PingResult.Ping from double to long

+++ MinecraftQuery/Models/PingResult.cs
@@ -5,7 +5,7 @@
         public string HostAddress { get; }
         public ushort HostPort { get; }
         public bool ServerAvailable { get; internal set; }
-        public double Ping { get; internal set; }
+        public long Ping { get; internal set; }
 
         internal PingResult(string hostAddress, ushort hostPort)
         {
